refractor: Rename the enjin api package to inapi(input api), this package provides interfaces and abstract classes that are implemented by apis, e.g the enjin api
