+++ src/main/kotlin/com/recraftedcivilizations/charactercards/inapi/IEndPoint.kt
@@ -1,4 +1,4 @@
-package com.recraftedcivilizations.charactercards.enjinapi
+package com.recraftedcivilizations.charactercards.inapi
 
 /**
  * Describe a class with an endpoint of some kind
