+++ src/test/kotlin/com/recraftedcivilizations/charactercards/inapi/RequestTest.kt
@@ -1,4 +1,4 @@
-package com.recraftedcivilizations.charactercards.enjinapi
+package com.recraftedcivilizations.charactercards.inapi
 
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
